only write if content is different

+++ lib/munge/core/write.rb
@@ -12,6 +12,8 @@ module Munge
 
         FileUtils.mkdir_p(File.dirname(abspath))
 
+        return if File.exist?(abspath) && File.read(abspath) == content
+
         File.write(abspath, content)
       end
